+++ Frameworks/EX2Kit/UIComponents/EX2InfinitePagingScrollView.m
@@ -115,10 +115,10 @@
     self.contentOffset = self.centerOffset;
     
     // Load the views for the visible pages (2 pages to the left, center page, and two pages to the right)
-    int start = self.currentPageIndex - 2;
+    long start = self.currentPageIndex - 2;
     start = start < 0 && !self.isWrapLeft ? 0 : start;
     
-    int end = self.currentPageIndex + 2;
+    long end = self.currentPageIndex + 2;
     if (!self.isWrapRight)
         end = end > self.numberOfPages - 1 ? self.numberOfPages - 1 : end;
     
@@ -131,7 +131,7 @@
     }
     else
     {
-        for (int i = start; i <= end; i++)
+        for (long i = start; i <= end; i++)
         {
             @autoreleasepool
             {
